Rename admin headers name
