+++ layouts/admin/ApplicationsLayout.tsx
@@ -1,11 +1,11 @@
-import AdminHeader from '../../components/AdminHeader'
+import ApplicationsHeader from '../../components/Admin/Headers/Applictions'
 import Footer from '../../components/Footer'
 import styles from '../../styles/Home.module.css'
 
-export default function AdminLayout({ children }) {
+export default function ApplicationsLayout({ children }) {
   return (
     <div className={styles.container}>
-      <AdminHeader />
+      <ApplicationsHeader />
       <div className="md:px-12">{children}</div>
       <Footer />
     </div>
