+++ pages/admin/data/awardees.tsx
@@ -5,7 +5,7 @@ import UpdateConstantModal from '../../../components/Admin/Modals/UpdateConstant
 import UpdateGroupModal from '../../../components/Admin/Modals/UpdateGroup'
 import requireAuth from '../../../components/requireAuth'
 import Roles from '../../../constants/roles'
-import AdminLayout from '../../../layouts/admin/admin-layout'
+import AdminLayout from '../../../layouts/admin/ApplicationsLayout'
 import { AwardeeSessionType, AwardeeType, AwardeesType } from '../../../types'
 import {
   deleteAwardeesSession,
