+++ pages/admin/interviewers/index.tsx
@@ -3,7 +3,7 @@ import SetDropdown from '../../../components/Admin/SetDropdown'
 import Loading from '../../../components/Loading'
 import requireAuth from '../../../components/requireAuth'
 import Roles from '../../../constants/roles'
-import AdminLayout from '../../../layouts/admin/admin-layout'
+import AdminLayout from '../../../layouts/admin/ApplicationsLayout'
 import { Users } from '../../../types'
 import { getUsersByRole } from '../../api/getUserDetails'
 import { updateUserSets } from '../../api/updateUserSets'
